[8.x][Data Usage] setup integration tests (#197112)
Adds serverless api integration tests and basic functional smoke test
for data usage plugin

Both are skipped in mki until `xpack.dataUsage.enabled` is enabled by
default in serverless

- skipped in MKI because we'll need to make sure real credentials are
being used via the `xpack.dataUsage.autoops*`
- we start a mock server at localhost:9000 and set that the config
(`xpack.dataUsage.autoops.api.url=http://localhost:9000'`) along with
fake credentials for the other `xpack.dataUsage.autoops*` values. If
we're not in MKI these values will be used and the mock server will
respond to the request at `http://localhost:9000`. If we are in MKI, the
real values and credentials should be set, otherwise it will fail as
these kibana config values in the tests are not passed into the MKI
environment.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>